refactor(canvas): extract skia_canvas_opacity.cpp (R2-3 fourth slice)
Phase 4 R2-3 follow-up. Pull the 308-line save_layer + filter
composition slice out of skia_canvas.cpp into a sibling TU. Covers:

- set_opacity(alpha) — per-draw alpha shim
- save_layer(x, y, w, h, alpha) — whole-subtree alpha
- save_layer_with_blend(x, y, w, h, alpha, blend_mode) — layer + blend
- save_layer_with_filters(x, y, w, h, filter_chain) — CSS filter chain
  composition (blur, brightness, contrast, drop-shadow, hue-rotate,
  invert, opacity, saturate, sepia, grayscale)
- save_backdrop_filter(x, y, w, h, blur_radius) — CSS backdrop-filter

- skia_canvas.cpp: 3,657 → 3,348 lines (-309 net).
- skia_canvas_opacity.cpp: 362 lines.

Tests:
- pulp-canvas builds clean on macOS arm64 (Debug, GPU + Skia).
- pulp-test-canvas: 1,550 assertions / 63 cases — pass.

R2-3 progress: mask + gradients + box-shadow + opacity (4 slices).
Remaining: GPU SDF, SkSL shaders, cg_canvas.mm mirror.
